Updated the CallTable to accept longs, added real-time per-query millisecond tracking in StructuredRetrieval

+++ branches/2.0/galagosearch-core/src/main/java/org/galagosearch/core/util/CallTable.java
@@ -14,7 +14,7 @@ public class CallTable {
     private static boolean on = true;
     private CallTable() {}
 
-    public static void increment(String counterName, int inc) {
+    public static void increment(String counterName, long inc) {
         if (on) {
             counts.adjustOrPutValue(counterName, inc, inc);
         }
